% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfret-functions.R
\name{bm_nll}
\alias{bm_nll}
\title{Negative loglikelihood for the Brownian model Xn.}
\usage{
bm_nll(mu, sigma, Xt, dt)
}
\arguments{
\item{mu}{(NOT USED) Scalar mean parameter (see \code{\link[=ou_sim]{ou_sim()}}).}

\item{sigma}{Scalar diffusion parameter (see \code{\link[=ou_sim]{ou_sim()}}).}

\item{Xt}{Vector of \code{n_obs} observations from the OU process.}

\item{dt}{Interobservation time.}
}
\value{
The scalar value of the loglikelihood \code{loglik(gamma, mu, sigma | Xt)}.
}
\description{
Negative loglikelihood for the Brownian model Xn.
}
\details{
For simplicity, the loglikelihood contribution of the first observation is ignored.
}
