% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfret-functions.R
\name{bm_sim}
\alias{bm_sim}
\title{Generate observations from a Brownian motion with drift at regular time intervals.}
\usage{
bm_sim(mu, sigma, dt, n_obs, x0 = 0)
}
\arguments{
\item{mu}{Scalar drift parameter (see \strong{Details}).}

\item{sigma}{Scalar diffusion parameter (see \strong{Details}).}

\item{dt}{Interobservation time.}

\item{n_obs}{Number of observations to generate.}

\item{x0}{Initial value of the process at time \code{t = 0}.  Default value is \code{x0 = 0}.}
}
\value{
A vector of \code{n_obs} observations of the process at times \verb{t = dt, 2*dt, ..., n_obs * dt}.
}
\description{
Generate observations from a Brownian motion with drift at regular time intervals.
}
\details{
Brownian motion with drift is a Gaussian Markov process with transition density\preformatted{X_s+t | X_s ~ N(X_s + mu * t, sigma^2 * t).
}
}
\author{
Martin Lysy \email{mlysy@uwaterloo.ca}
}
