% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfret-functions.R
\name{bm_y_nll}
\alias{bm_y_nll}
\title{Negative loglikelihood for the Brownian model Yn.}
\usage{
bm_y_nll(mu, sigma, b0, b1, Xt, Yt, dt)
}
\arguments{
\item{mu}{Scalar mean parameter (see \code{\link[=ou_sim]{ou_sim()}}).}

\item{sigma}{Scalar diffusion parameter (see \code{\link[=ou_sim]{ou_sim()}}).}

\item{Xt}{Vector of \code{n_obs} observations from the OU process.}

\item{Yt}{Vector of \code{n_obs} observations from the OU process.}

\item{dt}{Interobservation time.}

\item{gamma}{Scalar mean reversion parameter (see \code{\link[=ou_sim]{ou_sim()}}).}

\item{beta0}{Scaler intercept parameter for Yt Poisson distribution}

\item{beta1}{Scaler slope parameter for Yt Poisson distribution}
}
\value{
The scalar value of the loglikelihood \code{loglik(gamma, mu, sigma | Xt, Yt)}.
}
\description{
Negative loglikelihood for the Brownian model Yn.
}
