% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_start.R
\name{find_optim_omega}
\alias{find_optim_omega}
\title{find_optim_omega}
\usage{
find_optim_omega(omegas, n_obs, dt, Y, beta0, beta1, method = "Nelder-Mead")
}
\arguments{
\item{omegas}{potential omegas to be search over}

\item{n_obs}{number of observations}

\item{dt}{time interval}

\item{Y}{observered value the number of photons recorded}

\item{beta0}{Scalar Poisson parameter}

\item{beta1}{Scalar Poisson parameter}

\item{method}{optimization method, default Nelder-Mead}
}
\value{
the omega returns the minimum theta hat
}
\description{
find_optim_omega
}
