% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfret-functions.R
\name{ou_nll}
\alias{ou_nll}
\title{Negative loglikelihood for the Ornstein-Uhlenbeck model Xn.}
\usage{
ou_nll(gamma, mu, sigma, Xt, dt)
}
\arguments{
\item{gamma}{Scalar mean reversion parameter (see \code{\link[=ou_sim]{ou_sim()}}).}

\item{mu}{Scalar mean parameter (see \code{\link[=ou_sim]{ou_sim()}}).}

\item{sigma}{Scalar diffusion parameter (see \code{\link[=ou_sim]{ou_sim()}}).}

\item{Xt}{Vector of \code{n_obs} observations from the OU process.}

\item{dt}{Interobservation time.}
}
\value{
The scalar value of the loglikelihood \code{loglik(gamma, mu, sigma | Xt)}.
}
\description{
Negative loglikelihood for the Ornstein-Uhlenbeck model Xn.
}
\details{
For simplicity, the loglikelihood contribution of the first observation is ignored.
}
