% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfret-functions.R
\name{ou_y_nll}
\alias{ou_y_nll}
\title{Negative loglikelihood for the Ornstein-Uhlenbeck model Yn.}
\usage{
ou_y_nll(gamma, mu, sigma, b0, b1, Xt, Yt, dt)
}
\arguments{
\item{gamma}{Scalar mean reversion parameter (see \code{\link[=ou_sim]{ou_sim()}}).}

\item{mu}{Scalar mean parameter (see \code{\link[=ou_sim]{ou_sim()}}).}

\item{sigma}{Scalar diffusion parameter (see \code{\link[=ou_sim]{ou_sim()}}).}

\item{b0}{Scaler intercept parameter for Yt Poisson distribution}

\item{b1}{Scaler slope parameter for Yt Poisson distribution}

\item{Xt}{Vector of \code{n_obs} observations from the OU process.}

\item{Yt}{Vector of \code{n_obs} observations from the OU process.}

\item{dt}{Interobservation time.}
}
\value{
The scalar value of the loglikelihood \code{loglik(gamma, mu, sigma | Xt, Yt)}.
}
\description{
Negative loglikelihood for the Ornstein-Uhlenbeck model Yn.
}
