% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1_omega_tau_multiple.R
\name{sim_1}
\alias{sim_1}
\title{Simulation 1}
\usage{
sim_1(
  beta0 = 10,
  beta1 = 0.5,
  omega = exp(-1),
  mu = 10,
  tau = 1,
  dt = 1,
  n_obs = 99,
  n_dataset = 100,
  method = "BFGS",
  num_multistart = 5
)
}
\arguments{
\item{beta0}{The value of beta0.}

\item{beta1}{The value of beta1.}

\item{omega}{Parameter in Ornstein-Uhlenbeck process.}

\item{mu}{Scalar mean parameter.}

\item{tau}{Variance of donor acceptor distance.}

\item{dt}{Interobservation time.}

\item{n_obs}{Number of observations to generate.}

\item{n_dataset}{Number of dataset to simulate for each \verb{\\beta}.}
}
\value{
A vector of rmse: rmse for inference on each simulated dataset.
}
\description{
Simulation 1
}
