% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2.R
\name{sim_2}
\alias{sim_2}
\title{Simulation 2}
\usage{
sim_2(
  from,
  beta0 = 20,
  beta1 = 0.5,
  gamma = 1,
  mu = 10,
  sigma = 0.1,
  dt = 1,
  n_obs = 99,
  n_dataset = 100,
  multi_start = TRUE,
  method = "Nelder-Mead"
)
}
\arguments{
\item{from}{simulate X from which model ('ou' or 'bm')}

\item{beta0}{the value of beta0}

\item{beta1}{the value of beta1}

\item{gamma}{Scalar mean reversion parameter}

\item{mu}{Scalar mean parameter}

\item{sigma}{Scalar diffusion parameter}

\item{dt}{Interobservation time.}

\item{n_obs}{Number of observations to generate.}

\item{n_dataset}{Number of dataset to simulate for each \verb{\\beta}}
}
\value{
a vector of picked model
}
\description{
Simulation 2
}
