% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfret-functions.R
\name{y_sim}
\alias{y_sim}
\title{Simulate number of photons exchanged at different time}
\usage{
y_sim(X, beta0, beta1)
}
\arguments{
\item{X}{Vector of \code{n_obs} observations of donor-acceptor distance.}

\item{beta0}{Scaler intercept parameter for Yt Poisson distribution}

\item{beta1}{Scaler slope parameter for Yt Poisson distribution}
}
\value{
Vector of \code{n_obs} observation of photons exchanged at different time.
}
\description{
Simulate number of photons exchanged at different time
}
