% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfret-package.R
\docType{package}
\name{smfret-package}
\alias{smfret}
\alias{smfret-package}
\title{smfret: smFRET Parameter Inference}
\description{
Parameter Inference of molecular interactions using Brownian motion model and Ornstein-Uhlenbeck process model.
}
\author{
\strong{Maintainer}: Kar Mei Koo \email{kmkoo@uwaterloo.ca}

Authors:
\itemize{
  \item Wanxin Li \email{w328li@uwaterloo.ca}
  \item Jayden Luo \email{j57luo@uwaterloo.ca}
  \item Yi Xiang \email{y25xiang@uwaterloo.ca}
}

}
\keyword{internal}
